90_VideoChapterAgent/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
openai
pydub
ffmpeg-python
python-multipart
speechrecognition
# openai-whisper
